+++ indexer/src/main/scala/au/csiro/data61/magda/indexer/crawler/CrawlerApi.scala
@@ -2,7 +2,7 @@ package au.csiro.data61.magda.indexer.crawler
 
 import au.csiro.data61.magda.indexer.search.SearchIndexer
 import au.csiro.data61.magda.api.BaseMagdaApi
-import au.csiro.data61.magda.model.Registry.RecordsChangedWebHookPayload
+import au.csiro.data61.magda.model.Registry.WebHookPayload
 import akka.event.LoggingAdapter
 import akka.http.scaladsl.server.Directives._
 import scala.util.Failure
