+++ indexer/src/main/scala/au/csiro/data61/magda/indexer/external/csw/GMDCSWImplementation.scala
@@ -12,7 +12,7 @@ import akka.http.scaladsl.server.Directives._
 import akka.http.scaladsl.unmarshalling.Unmarshal
 import akka.stream.Materializer
 import scala.concurrent.ExecutionContext
-import au.csiro.data61.magda.model.temporal._
+import au.csiro.data61.magda.model.Temporal._
 import au.csiro.data61.magda.model.misc._
 import au.csiro.data61.magda.util.DateParser._
 import com.monsanto.labs.mwundo.GeoJson.Coordinate
