+++ int-test/src/test/scala/au/csiro/data61/magda/test/util/MagdaGeneratorTest.scala
@@ -5,7 +5,7 @@ import org.scalactic.anyvals.PosInt
 import org.scalatest.prop.GeneratorDrivenPropertyChecks
 
 trait MagdaGeneratorTest extends GeneratorDrivenPropertyChecks {
-  val processors = Math.max(2, if (ContinuousIntegration.isCi) Runtime.getRuntime.availableProcessors - 1 else Runtime.getRuntime.availableProcessors - 1)
+  val processors = Math.max(2, if (ContinuousIntegration.isCi) Runtime.getRuntime.availableProcessors / 2 else Runtime.getRuntime.availableProcessors - 1)
   val minSuccessful = if (ContinuousIntegration.isCi) 100 else 20
   implicit override val generatorDrivenConfig: PropertyCheckConfiguration =
     PropertyCheckConfiguration(workers = PosInt.from(processors).get, sizeRange = if (ContinuousIntegration.isCi) 25 else 50, minSuccessful = PosInt.from(minSuccessful).get)
