+++ registry-api/src/main/scala/au/csiro/data61/magda/registry/AspectsService.scala
@@ -11,6 +11,7 @@ import akka.http.scaladsl.model.StatusCodes
 import io.swagger.annotations._
 import spray.json._
 import gnieh.diffson.sprayJson._
+import au.csiro.data61.magda.model.Registry._
 
 import scala.util.{Failure, Success}
