+++ registry-api/src/test/scala/au/csiro/data61/magda/registry/AspectsServiceSpec.scala
@@ -4,6 +4,7 @@ import akka.http.scaladsl.model.StatusCodes
 import gnieh.diffson._
 import gnieh.diffson.sprayJson._
 import spray.json._
+import au.csiro.data61.magda.model.Registry._
 
 class AspectsServiceSpec extends ApiSpec {
   describe("GET") {
