+++ registry-api/src/test/scala/au/csiro/data61/magda/registry/RecordsServiceSpec.scala
@@ -6,6 +6,7 @@ import gnieh.diffson.sprayJson._
 import spray.json._
 import au.csiro.data61.magda.model.Registry.Record
 import scala.concurrent.duration._
+import au.csiro.data61.magda.model.Registry._
 
 class RecordsServiceSpec extends ApiSpec {
   describe("GET") {
